refactor(sql-editor): extract AI/diff + shortcuts hooks (decompose 5/6)
Extract the Assistant/diff cluster into useSqlEditorAi (SQL completion, the
ask-AI prompt flow, accept/discard diff handlers, debug-prompt helpers, and
the diff lifecycle effects) and the keyboard/shortcut wiring into
useSqlEditorShortcuts. SQLEditorContent now composes these hooks.

Behavior-preserving: the extracted function bodies, effect logic, dependency
arrays, and JSX are unchanged from the previous monolith. The untrusted->safe
SQL promotion (acceptUntrustedSql) continues to happen in the run/explain
gesture and warning-modal handlers, as close to the user action as possible.

The Phase-1 characterization suite remains green.
